fix: problem with hidden text on small or zoomed in screens
