+++ src/components/Section.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import styled, { DefaultTheme, ThemeProps } from 'styled-components';
 import tw from 'twin.macro';
-import { sizes } from '../helpers';
-import { useViewportSize } from '../hooks';
+import { useSectionHeight } from '../hooks';
 
 type StyleProps = {
 	height?: number;
@@ -27,23 +26,10 @@ type Props = {
 };
 
 export default ({ children, headerSpace, ...props }: Props) => {
-	const [height, width] = useViewportSize();
-
-	const isMobile = () => width < sizes().laptop;
-	const getHeight = () => {
-		// Don't set a height if it's a mobile phone
-		// since we don't use a left menu on mobile.
-		if (isMobile()) {
-			return undefined;
-		}
-		if (height < sizes().minimumHeight) {
-			return sizes().minimumHeight - (headerSpace ? sizes().headerHeight : 0);
-		}
-		return headerSpace ? height - sizes().headerHeight : height;
-	};
+	const sectionHeight = useSectionHeight(headerSpace);
 
 	return (
-		<Section height={getHeight()} {...props}>
+		<Section height={sectionHeight} {...props}>
 			{children}
 		</Section>
 	);
