+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useViewportSize } from './useViewportSize';
 export { default as useInterval } from './useInterval';
+export { default as useSectionHeight } from './useSectionHeight';
